requirements.txt

python-dotenv
textblob
spacy
pandas
matplotlib
streamlit
google-api-python-client
scikit-learn
